Move policy commands under policy category
