+++ tests/integration-phase2.test.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { DuckDBStore } from "../src/cli/storage/duckdb.js";
-import { runQuarantine } from "../src/cli/commands/quarantine.js";
+import { runQuarantine } from "../src/cli/commands/policy/quarantine.js";
 import { runBisect } from "../src/cli/commands/debug/bisect.js";
 import { runSample } from "../src/cli/commands/sample.js";
 import { SimpleResolver } from "../src/cli/resolvers/simple.js";
